Facilitate easier reuse of a ByteBuffer object by allowing it to wrap a different byte array in its Reset() function.
This is useful in cases where the client frequently receives
serialized data from an event that supplies a byte[]. The client
shouldn't make any assumptions about the byte array it's receiving,
as the event could be supplying a different byte[] for subsequent
invocations. To keep object allocation and garbage collection at
a minimum, it's preferable to reuse an existing ByteBuffer object
for each event instead of allocating a new one every time.